+++ src/pages/requests/requisition-requests/view-requisition-request/index.tsx
@@ -252,7 +252,7 @@ export default function ViewRequisitionRequest() {
         open={openImage}
         setOpen={setOpenImage}
         showXicon={true}
-        title="View Reqeust Images"
+        title="View Request Image(s)"
         className=" max-w-screen-sm"
       >
         <ImageViewer