+++ src/types/apiData/referrals.ts
@@ -1,3 +1,5 @@
+import { apartmentById } from "./apartment";
+import { bookingsById } from "./bookings";
 import { customersById } from "./customers";
 
 export interface referrals {
@@ -5,9 +7,13 @@ export interface referrals {
   user_id: number;
   referred_user: customersById;
   referral_code: string;
+  booking_id: number;
+  shortlet_id: number;
   is_deleted: number;
   deleted_at: null;
   created_at: string;
   updated_at: string;
   user: customersById;
+  booking: bookingsById;
+  shortlet: apartmentById;
 }